[Refactor] #22 Replace classnames with clsx and change the fontawesome icon to svg tag
