--- components/DetailHeader/DetailHeader.module.scss
@@ -1,22 +1,0 @@
-.ly_header_nav {
-  display: flex;
-  background: #fff;
-  justify-content: space-around;
-  align-items: center;
-  width: 100%;
-  height: 30px;
-  border: 1px solid #dee2e6;
-  margin-bottom: 5px;
-
-  padding: 20px;
-}
-
-.ly_header_nav_cat {
-  text-decoration: none;
-  color: #adb5bd;
-  font-size: 12px;
-}
-
-.ly_header_nav_cat > .active {
-  color: black;
-}
