--- components/StockInfoList/StockInfoList.module.scss
@@ -1,8 +1,0 @@
-.bl_stockInfoList {
-  li {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    border-bottom: 1px solid #ddd;
-    padding: 6px 10px;
-  }
-}
