+++ package.json
@@ -9,8 +9,13 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "classnames": "^2.3.1",
+    "@fortawesome/fontawesome-svg-core": "^1.3.0",
+    "@fortawesome/free-brands-svg-icons": "^6.0.0",
+    "@fortawesome/free-regular-svg-icons": "^6.0.0",
+    "@fortawesome/free-solid-svg-icons": "^6.0.0",
+    "@fortawesome/react-fontawesome": "^0.1.17",
     "@types/node": "^17.0.14",
+    "clsx": "^1.1.1",
     "eslint-config-prettier": "^8.3.0",
     "husky": "^7.0.4",
     "lint-staged": "^12.2.2",
