+++ pages/_app.tsx
@@ -6,7 +6,7 @@ import Head from 'next/head';
 import '../styles/globals.scss';
 
 import Spinner from '../components/ui/spinner';
-import NavBar from '../components/NavBar/NavBar';
+import NavBar from '../components/NavBar';
 
 function MyApp({ Component, pageProps }: AppProps) {
   const [isLoading, setIsLoading] = useState(false);
